pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modelpilot"
version = "0.0.1"
description = "Generic, rigorous model tuning with evaluation, reporting, and an agentic reviewer."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name="Your Name", email="you@example.com"}]
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "jinja2>=3.1",
  "mlflow>=2.10.0",
  "faiss-cpu; platform_system != 'Windows'",
  "faiss-windows; platform_system == 'Windows'",
  "streamlit>=1.34",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "black",
  "pytest",
  "pytest-cov"
]
extras = [
  "optuna>=3.6.0",
  "ngboost>=0.5.1",
  "shap>=0.44.0"
]

[project.scripts]
modelpilot = "modelpilot.cli.main:app"
